Move `build_and_save_index` method to `NmslibIndexBuilderMixin`
